Add related case studies for each existing case study
