+++ src/templates/industry.tsx
@@ -4,45 +4,8 @@ import "../pages/styles.scss";
 import classNames from "classnames";
 import { ContactFormSection } from "../components/ContactFormSection";
 import { SEO } from "../components/SEO";
-import { useCaseStudies } from "../hooks/useCaseStudies";
+import { CaseStudies } from "../components/CaseStudies";
 
-const CaseStudies = ({ title, ids }: { title: string; ids: string[] }) => {
-  const baseStudies = useCaseStudies();
-
-  const studies = ids.map((id) => {
-    const caseStudy = baseStudies.find((cs) => cs.caseStudyId === id);
-    if (!caseStudy) {
-      throw new Error(`Case study with id ${id} not found`);
-    }
-    return caseStudy;
-  });
-
-  return (
-    <section className="industry-case-studies">
-      <h3>{title}</h3>
-      <div className="industry-case-studies__list">
-        {studies.map((study) => (
-          <a
-            href={study.path}
-            className="industry-case-studies__case-study"
-            key={study.caseStudyId}
-          >
-            <div className="industry-case-studies__case-study__logo">
-              <img src={study.logo} alt={study.productName} />
-            </div>
-            <div className="industry-case-studies__case-study__content">
-              <div className="industry-case-studies__case-study__content__title">
-                <h4>{study.productName}</h4>
-                <div className="card__link-button"></div>
-              </div>
-              <p>{study.productDescription}</p>
-            </div>
-          </a>
-        ))}
-      </div>
-    </section>
-  );
-};
 
 const Industry = (props: any) => {
   const { pageContext } = props;
